client socket

+++ src/Common/CustomAxios.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import {RESPONSE_FORBIDDEN, RESPONSE_UNAUTHORIZED} from "./Response";
 
 export const customAxios = axios.create({
     baseURL: "http://localhost:8080"
@@ -8,7 +9,7 @@ customAxios.interceptors.request.use(
     function (config)
     {
         config.headers.ContentType = "application/json; charset=utf-8";
-        //config.headers.authorization = localStorage.getItem("jwt");
+        config.headers.authorization = localStorage.getItem("jwt");
         return config;
     }
 )
@@ -16,26 +17,31 @@ customAxios.interceptors.request.use(
 customAxios.interceptors.response.use(
     function (response)
     {
-        /*if(response.headers['authorization'] !== undefined)
+        if(response.headers['authorization'] !== undefined)
         {
             localStorage.setItem("jwt", response.headers['authorization']);
-            if(response.headers['refresh'] !== undefined)
-            {
-                localStorage.setItem("refresh", response.headers['refresh']);
-            }
-        }*/
+        }
         return response;
     },
     function (error) {
-        /*if (error.response.status === 401 || error.response.status === 403) {
-            axios.post(`http://localhost:8080/logout`, null, {
-                headers:
-                    {
-                        "Content-Type": "application/json; charset=utf-8",
-                        "authorization": localStorage.getItem("jwt")}
-            }).then();
-            localStorage.clear();
-        }*/
+        if(error.response.headers['authorization'] !== undefined)
+        {
+            localStorage.setItem("jwt", error.response.headers['authorization']);
+        }
+        if (error.response.request.status === RESPONSE_UNAUTHORIZED)
+        {
+            alert("로그인 해주세요");
+            axios.post(`http://localhost:8080/logout`)
+                .then((response)=>{
+                    localStorage.clear();
+                }).catch((error)=>{
+                    alert(error.response.request.status);
+            });
+        }
+        else if(error.response.request.status === RESPONSE_FORBIDDEN)
+        {
+            alert("권한이 없습니다");
+        }
         return Promise.reject(error);
     }
 );
